moved function `generate_diff` into a separate module
