+++ gendiff/__init__.py
@@ -1,6 +1,6 @@
 """Gendiff project."""
 
-from gendiff.scripts.gendiff import generate_diff
+from gendiff.generate_diff import generate_diff
 
 __all__ = (  # noqa: WPS410
     'generate_diff',
