+++ gendiff/scripts/gendiff.py
@@ -3,26 +3,8 @@
 """gendiff package."""
 
 
-from gendiff.find_diff import find_diff
-from gendiff.format_diff import format_diff
-from gendiff.loader import loader
 from gendiff.args_parser import args_parse
-
-
-def generate_diff(file1, file2, formater='stylish'):
-    """Func generate diff of two files.
-
-    Args:
-        file1: path to file1
-        file2: path to file2
-        formater: format for output
-
-    Returns:
-        string with diff
-    """
-    content1, content2 = loader(file1), loader(file2)
-    diff = find_diff(content1, content2)
-    return format_diff(diff, formater)
+from gendiff.generate_diff import generate_diff
 
 
 def main():
